travel_planning_system/hotel_booking_agent_crewai/pyproject.toml
[project]
name = "hotel-booking-agent-crewai"
version = "0.1.0"
description = "Hotel Booking Agent using CrewAI for hotel research and booking."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "crewai>=0.70.0",
    "python-dotenv",
    "requests",
    "fastapi",
    "uvicorn",
    "pydantic",
    "groq",
    "langchain-groq",
    "crewai-tools[mcp]>=0.75.0",
] 
